+++ R/print.mdl_rn.R
@@ -79,7 +79,7 @@ print.mdl_rn <- function(x, timestep = x$max_i, digits = 3, ...) {
   save_time <- round(x$save_each * x$min_per_i / 60 / 24, digits = 2)
 
   # print result
-  cat(paste0("Total time : ", i, " iterations (", total_time, " days) [Burn-in: ", x$burn_in * 100, "%]\n",
+  cat(paste0("Total time : ", i, " iterations (", total_time, " days) [Burn-in: ", x$burn_in, " iter.]\n",
              "Saved each : ", x$save_each, " iterations (", save_time, " days)\n",
              "Seafloor   : ", x$extent, ", ", nrow(x$coords_reef), " reef cells\n",
              "Fishpop    : ", x$starting_values$pop_n, " indiv (reef_attraction: ", x$reef_attraction, ")\n",