+++ R/summarize_mdlrn.R
@@ -23,9 +23,6 @@ summarize_mdlrn <- function(result, summary = c("min", "mean", "max")) {
   # get timesteps
   timestep_seafloor <- result$seafloor$timestep
 
-  # calculate iteration of burn_in
-  burn_in_itr <- result$max_i * result$burn_in
-
   # get cols to summarise
   seafloor <- subset(result$seafloor, select = c("ag_biomass", "bg_biomass",
                                                  "nutrients_pool", "detritus_pool",
@@ -44,7 +41,7 @@ summarize_mdlrn <- function(result, summary = c("min", "mean", "max")) {
                           each = nrow(seafloor) / length(summary))
 
   # add burn_in col
-  seafloor$burn_in <- ifelse(test = seafloor$timestep < burn_in_itr,
+  seafloor$burn_in <- ifelse(test = seafloor$timestep < result$burn_in,
                              yes = "yes", no = "no")
 
   if (nrow(result$fishpop > 0)) {
@@ -69,7 +66,7 @@ summarize_mdlrn <- function(result, summary = c("min", "mean", "max")) {
                            each = nrow(fishpop) / length(summary))
 
     # add burn_in col
-    fishpop$burn_in <- ifelse(test = fishpop$timestep < burn_in_itr,
+    fishpop$burn_in <- ifelse(test = fishpop$timestep < result$burn_in,
                               yes = "yes", no = "no")
 
   # no fish present